refactor: PROD_LIST => PRODUCTS, updateSelOpts => updateProductOptions, 함수분리
